+++ lab_1/VolodyaDudchak/head/add.h
@@ -1,0 +1,13 @@
+#include <bits/stdc++.h>
+using namespace std;
+
+class adres{
+    public:
+    int zip;
+    string city;
+    string str;
+    adres();
+    adres(int zp, string ct, string st);
+    void inp();
+    void out();
+};
