+++ lab_1/VolodyaDudchak/head/zal.h
@@ -1,0 +1,14 @@
+#include <bits/stdc++.h>
+#include "../sours/pred.cpp"
+using namespace std;
+
+class zalikov{
+    public:
+    int name;
+    predmet* preds;
+    int sz;
+    zalikov();
+    zalikov(int zp, int ct, predmet* pd);
+    void inp();
+    void out();
+};
