+++ lab_1/VolodyaDudchak/sours/in.txt
