+++ lab_1/VolodyaDudchak/sours/main.cpp
