+++ lab_1/VolodyaDudchak/sours/out.txt
