+++ lab_1/VolodyaDudchak/sours/zal.cpp
@@ -1,31 +1,24 @@
 #include <bits/stdc++.h>
-#include "pred.h"
-using namespace std;
+#include "../head/zal.h"
 
-class zalikov{
-    public:
-    int name;
-    predmet* preds;
-    int sz;
-    zalikov(){
+    zalikov::zalikov(){
         sz = 0, name = 0, preds = new predmet[0];
     }
-    zalikov(int zp, int ct, predmet* pd)
+    zalikov::zalikov(int zp, int ct, predmet* pd)
     {
         sz = zp, name = ct, preds = pd;
     }
-    void inp()
+    void zalikov::inp()
     {
         in >> name>> sz;
         preds = new predmet[sz];
         for(int i = 0; i < sz;i++)
             preds[i].inp();
     }
-    void out()
+    void zalikov::out()
     {
         cout << "Zalick: ";
         cout << name << ' ' << sz << ' ' << endl;
         for(int i = 0; i < sz;i++)
             preds[i].out();
     }
-};
